requirements.txt
google_auth_oauthlib==0.4.3
requests==2.25.1
beautifulsoup4==4.9.3
google_api_python_client==2.0.2
protobuf==3.15.6
lxml==4.6.2
